Refactor: add new FileOperator class to deal with file operations in handlers.
