+++ src/httpserver/handlers/Handler.java
@@ -1,19 +1,20 @@
 package httpserver.handlers;
 
+import httpserver.utilities.FileOperator;
 import httpserver.utilities.FileTypeDecoder;
 import httpserver.utilities.Method;
 import httpserver.request.Request;
 import httpserver.utilities.FileContentConverter;
 import httpserver.response.Response;
 
-import java.io.File;
 import java.io.IOException;
-import java.nio.file.Files;
-import java.nio.file.Paths;
 import java.util.ArrayList;
 
 public abstract class Handler {
 
+    private static final FileContentConverter fileContentConverter = new FileContentConverter();
+    private static final FileTypeDecoder fileTypeDecoder = new FileTypeDecoder();
+    private static final FileOperator fileOperator = new FileOperator();
     private final ArrayList<Method> handledMethods = new ArrayList<>();
     private HandlerType typeOfHandler = null;
 
@@ -34,22 +35,18 @@ public abstract class Handler {
     }
 
     public FileContentConverter getFileContentConverter() {
-        return new FileContentConverter();
+        return fileContentConverter;
     }
 
     public FileTypeDecoder getFileTypeDecoder() {
-        return new FileTypeDecoder();
+        return fileTypeDecoder;
     }
 
-    public File getRequestedFile(Request request, String rootPath) {
-        return new File(rootPath + "/" + request.getPath());
+    public FileOperator getFileOperator() {
+        return fileOperator;
     }
 
     public boolean handles(Request request) {
         return (this.handledMethods.contains(request.getMethod()) && this.coversPathFromRequest(request));
-    }
-
-    public boolean fileExistsOnPath(Request request, String rootPath) {
-        return Files.exists(Paths.get(rootPath + request.getPath()));
     }
 }
